[IMP] mrp_repair_component_history (Generation Logic Update)
Odoo's migration script splits the manufacturing orders into individuals for serials. In many older manufacturing orders, this split consumed all the components on the main MO and it generated the MO without components. For this scenario we will reference the BoM for the component history.

There is a ticket with Odoo for this issue, whether they fix their script or not is undetermined at this time.

https://pm.opensourceintegrators.com/web#cids=1&menu_id=218&action=334&active_id=1&model=helpdesk.ticket&view_type=form&id=67747